+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Add `app.tools.go` tool for calling `go` commands with the proper version
 - Add Git author details to telemetry
+- Improve telemetry for the `inv` command
 
 ## 0.13.1 - 2025-05-28