+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## 0.14.0 - 2025-05-30
+
 ***Added:***
 
 - Add `app.tools.go` tool for calling `go` commands with the proper version